+++ demo-service/src/main/java/app/DemoApp.java
@@ -3,7 +3,7 @@ package app;
 import app.domain.Product;
 import app.service.ProductService;
 import core.framework.api.AbstractApplication;
-import core.framework.api.module.StandardAppModule;
+import core.framework.api.module.SystemModule;
 import core.framework.api.util.ClasspathResources;
 import core.framework.api.util.YAML;
 
@@ -15,15 +15,12 @@ import java.time.Duration;
 public class DemoApp extends AbstractApplication {
     @Override
     protected void initialize() {
-        load(new StandardAppModule("app.properties"));
+        load(new SystemModule("sys.properties"));
 
         cache().add(Product.class, Duration.ofSeconds(60));
         cache().add("product-name-id", Integer.class, Duration.ofSeconds(60));
 
         bind(YAML.load(BusinessConfig.class, ClasspathResources.text("config.yml")));
-
-        db().url("jdbc:mysql:192.168.2.2:3306/main")
-            .user("root");
 
         db().repository(Product.class);
         bind(ProductService.class);