+++ demo-service/src/main/java/app/service/ProductService.java
@@ -3,17 +3,10 @@ package app.service;
 import app.domain.Product;
 import core.framework.api.cache.Cache;
 import core.framework.api.db.Database;
-import core.framework.api.util.Lists;
-import core.framework.api.util.Maps;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import javax.inject.Inject;
 import javax.inject.Named;
-import java.lang.reflect.Array;
 import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Map;
 
 /**
  * @author neo