+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 3.1.9 (7/8/2015)
 * fix api webservice client encode path param
+* renamed StandardAppModule to SystemModule, and added jdbc pool properties
 
 ### 3.1.8 (6/30/2015)
 * fix requestURL(), parse x-forwarded-port to get requested port